pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ndlocr-lite-server"
version = "1.0.0"
description = "NDLOCR-Lite Web OCR Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lxml==5.4.0",
    "networkx==3.3",
    "onnxruntime-gpu==1.23.2",
    "pillow==12.1.1",
    "ordered-set==4.1.0",
    "protobuf==6.31.1",
    "PyYAML==6.0.1",
    "tqdm==4.66.4",
    "pypdfium2==4.30.0",
    "numpy==2.2.2",
    "fastapi==0.111.0",
    "uvicorn==0.29.0",
    "python-dotenv==1.0.1"
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["ocr", "deim", "parseq", "ndl_parser", "tablerecog"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"model" = ["*.onnx"]
"config" = ["*.yaml"]
